improve checks and logging #97
